random word generates and turns int underscores, set styling for words and imgs
