+++ game.html
@@ -15,11 +15,9 @@
         <img src="./assests/img/hangman-title.png" alt="">
         <h1 id="welcomeUser"></h1>
     </header>
-    <section class="word_container">
-
-    </section>
+    <section class="word_container"></section>
     <section class="img_container">
-
+        <img src="./assests/img/h-5.jpg" alt="">
     </section>
     <section class="btn-container"></section>
 </body>
