+++ scripts/modules/username.js
@@ -1,4 +1,4 @@
-function setUsername () {
+export function setUsername () {
     const username = localStorage.getItem("username");
 
     if (username) {
@@ -7,5 +7,13 @@ function setUsername () {
         window.location.href = "index.html";
     }
 }
-    
-export default setUsername
+
+export function getUsername() {
+    document.getElementById("usernameForm").addEventListener("submit", function(event) {
+        event.preventDefault(); 
+        const username = document.getElementById("username").value;
+        localStorage.setItem("username", username);
+        window.location.href = "game.html";
+    });
+}
+
