+++ scss/sections/_buttons.scss
@@ -2,9 +2,9 @@
 @use "../particals/mixins";
 
 .btn-container {
-    width: 65%;
+    width: 60%;
     margin: auto;
-    @include mixins.flex(row);
+    @include mixins.flex(row, space-evenly);
     flex-wrap: wrap;
     #btn {
         @include mixins.btn-style;
