+++ styles.css
@@ -77,12 +77,31 @@ form button:hover {
   padding: 10px 20px;
 }
 
+.word_container {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  font-size: 2.5rem;
+}
+
+.img_container {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  margin: 20px;
+}
+.img_container img {
+  width: 250px;
+}
+
 .btn-container {
-  width: 65%;
+  width: 60%;
   margin: auto;
   display: flex;
   flex-direction: row;
-  justify-content: center;
+  justify-content: space-evenly;
   align-items: center;
   flex-wrap: wrap;
 }
