+++ styles.scss
@@ -1,6 +1,8 @@
 @use "./scss/sections/header";
 @use "./scss/sections/start-game";
 @use "./scss/sections/game-header";
+@use "./scss/sections/word";
+@use "./scss/sections/image";
 @use "./scss/sections/buttons";
 @use "./scss/particals/colors" as c;
 @use "./scss/particals/typography" as font;
