+++ words/example-words.json
@@ -1,0 +1,202 @@
+[
+  "apple",
+  "banana",
+  "orange",
+  "grape",
+  "kiwi",
+  "pear",
+  "peach",
+  "plum",
+  "melon",
+  "lemon",
+  "pineapple",
+  "mango",
+  "papaya",
+  "coconut",
+  "strawberry",
+  "blueberry",
+  "raspberry",
+  "blackberry",
+  "cherry",
+  "apricot",
+  "tomato",
+  "carrot",
+  "potato",
+  "onion",
+  "garlic",
+  "pepper",
+  "lettuce",
+  "broccoli",
+  "spinach",
+  "zucchini",
+  "cucumber",
+  "celery",
+  "cauliflower",
+  "asparagus",
+  "mushroom",
+  "pumpkin",
+  "radish",
+  "beetroot",
+  "turnip",
+  "parsnip",
+  "elephant",
+  "giraffe",
+  "tiger",
+  "lion",
+  "cheetah",
+  "leopard",
+  "zebra",
+  "rhino",
+  "hippo",
+  "buffalo",
+  "kangaroo",
+  "koala",
+  "panda",
+  "sloth",
+  "chimpanzee",
+  "gorilla",
+  "orangutan",
+  "lemur",
+  "meerkat",
+  "otter",
+  "shark",
+  "whale",
+  "dolphin",
+  "seal",
+  "octopus",
+  "jellyfish",
+  "lobster",
+  "crab",
+  "shrimp",
+  "starfish",
+  "eagle",
+  "sparrow",
+  "parrot",
+  "penguin",
+  "ostrich",
+  "flamingo",
+  "peacock",
+  "hummingbird",
+  "owl",
+  "falcon",
+  "house",
+  "apartment",
+  "mansion",
+  "cottage",
+  "bungalow",
+  "castle",
+  "villa",
+  "shack",
+  "chalet",
+  "palace",
+  "bed",
+  "chair",
+  "table",
+  "desk",
+  "sofa",
+  "couch",
+  "cabinet",
+  "wardrobe",
+  "bookshelf",
+  "dresser",
+  "car",
+  "bicycle",
+  "motorcycle",
+  "scooter",
+  "truck",
+  "bus",
+  "train",
+  "airplane",
+  "helicopter",
+  "boat",
+  "submarine",
+  "rocket",
+  "spaceship",
+  "hovercraft",
+  "tram",
+  "trolley",
+  "taxi",
+  "ferry",
+  "yacht",
+  "canoe",
+  "violin",
+  "guitar",
+  "piano",
+  "trumpet",
+  "flute",
+  "drums",
+  "clarinet",
+  "saxophone",
+  "cello",
+  "harp",
+  "concert",
+  "symphony",
+  "melody",
+  "harmony",
+  "rhythm",
+  "tempo",
+  "note",
+  "scale",
+  "chord",
+  "tune",
+  "dog",
+  "cat",
+  "rabbit",
+  "hamster",
+  "goldfish",
+  "parakeet",
+  "ferret",
+  "guinea",
+  "chinchilla",
+  "gerbil",
+  "winter",
+  "spring",
+  "summer",
+  "autumn",
+  "snow",
+  "rain",
+  "hail",
+  "sleet",
+  "fog",
+  "storm",
+  "mountain",
+  "river",
+  "lake",
+  "ocean",
+  "forest",
+  "desert",
+  "valley",
+  "canyon",
+  "island",
+  "waterfall",
+  "kitchen",
+  "bathroom",
+  "bedroom",
+  "livingroom",
+  "garage",
+  "basement",
+  "attic",
+  "hallway",
+  "balcony",
+  "patio",
+  "science",
+  "history",
+  "math",
+  "geography",
+  "chemistry",
+  "biology",
+  "physics",
+  "astronomy",
+  "literature",
+  "philosophy",
+  "football",
+  "basketball",
+  "tennis",
+  "golf",
+  "soccer",
+  "baseball",
+  "hockey",
+  "cricket",
+  "rugby",
+  "volleyball"
+]
